Add conflict for too old jms/serializer versions
The supported `jms/serializer` version of the metadata parser is `> 2.3`, but this is only reflected in the `require-dev` section and will not be considered when installing the library via composer somewhere else.
Adding a `conflict` with versions lower than `2.3` fixes that issue and makes sure, that only a supported version will be installed.