Update `php-cs-fixer` to version `3.68.1`
With the latest version we can ensure, that typed properties which also have a default `null` value will also be marked explicitly as nullable.
E.g.

```php
function foo(string $string = null) {}
```

becomes

```php
function foo(?string $string = null) {}
```

This is necessary, as PHP 8.4 will otherwise trigger a deprecation message if the type is not marked properly as nullable.
Some unnecessary rules that come with this new version were also deactivated, as they do not make a lot of sense in this project.