Add test for using camelCase serialized name
There is currently a bug where when you use a camelCase name for the `SerializedName` attribute which would result in the same snake_case name, that the `RawClassMetadata::renameProperty` method would throw an exception because it thinks we are trying to use the same name again.
This test only acts as a reference so we can fix the actual bug.