Decouple property naming from `PropertyCollection`
With this change it is now possible to pass different property naming strategies to the different parsers.
This also provides more freedom when it comes to property collection handling, as you don't have to be aware that the collection itself creates the serialized name, but instead it is up to the caller to decide what serialized name should be used.